applying inheritance in all classes
