+++ src/main/java/com/sensezzaria/entidades/Alimento.java
@@ -1,0 +1,22 @@
+package com.sensezzaria.entidades;
+
+public class Alimento {
+    private Long id;
+    private Float valor;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Float getValor() {
+        return valor;
+    }
+
+    public void setValor(Float valor) {
+        this.valor = valor;
+    }
+}
