+++ src/main/java/com/sensezzaria/entidades/Pizza.java
@@ -1,18 +1,8 @@
 package com.sensezzaria.entidades;
 
-public class Pizza {
-     private Long id;
+public class Pizza extends Alimento{
      private String sabor;
      private Integer tamanho;
-     private Float valor;
-
-     public Long getId() {
-          return id;
-     }
-
-     public void setId(Long id) {
-          this.id = id;
-     }
 
      public String getSabor() {
           return sabor;
@@ -30,11 +20,4 @@ public class Pizza {
           this.tamanho = tamanho;
      }
 
-     public Float getValor() {
-          return valor;
-     }
-
-     public void setValor(Float valor) {
-          this.valor = valor;
-     }
 }
