+++ src/main/java/com/sensezzaria/entidades/Sobremesa.java
@@ -1,31 +1,12 @@
 package com.sensezzaria.entidades;
 
-public class Sobremesa {
-    private Long id;
+public class Sobremesa extends Alimento{
     private String nome;
-    private Float valor;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
     public String getNome() {
         return nome;
     }
 
     public void setNome(String nome) {
         this.nome = nome;
-    }
-
-    public Float getValor() {
-        return valor;
-    }
-
-    public void setValor(Float valor) {
-        this.valor = valor;
     }
 }
